Package rename to argument
