+++ leshan-client-core/src/main/java/org/eclipse/leshan/client/object/LwM2mTestObject.java
@@ -25,8 +25,8 @@ import org.eclipse.leshan.client.resource.SimpleInstanceEnabler;
 import org.eclipse.leshan.client.servers.ServerIdentity;
 import org.eclipse.leshan.core.node.LwM2mResourceInstance;
 import org.eclipse.leshan.core.node.ObjectLink;
-import org.eclipse.leshan.core.request.execute.Argument;
-import org.eclipse.leshan.core.request.execute.Arguments;
+import org.eclipse.leshan.core.request.argument.Argument;
+import org.eclipse.leshan.core.request.argument.Arguments;
 import org.eclipse.leshan.core.response.ExecuteResponse;
 import org.eclipse.leshan.core.util.Hex;
 import org.eclipse.leshan.core.util.RandomStringUtils;
