+++ leshan-core/src/main/java/org/eclipse/leshan/core/request/argument/InvalidArgumentException.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Orange - Add better support for Arguments of Execute Operation.
  *******************************************************************************/
-package org.eclipse.leshan.core.request.execute;
+package org.eclipse.leshan.core.request.argument;
 
 /**
  * Exception thrown in case of invalid argument format while creating one.
