+++ leshan-core/src/test/java/org/eclipse/leshan/core/request/argument/ArgumentValidationTest.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Orange - Add better support for Arguments of Execute Operation.
  *******************************************************************************/
-package org.eclipse.leshan.core.request.execute;
+package org.eclipse.leshan.core.request.argument;
 
 import static org.junit.Assert.assertThrows;
